Protobuf: Encode UUID in binary. Add question to the response msg
Also:
- merge the dnsdist and recursor Remote Logger implementations.
- fix the protobuf logging over TCP in dnsdist
