+++ pdns/dnsdist-lua2.cc
@@ -582,8 +582,8 @@ void moreLua(bool client)
     g_lua.writeFunction("RemoteLogResponseAction", [](std::shared_ptr<RemoteLogger> logger) {
         return std::shared_ptr<DNSResponseAction>(new RemoteLogResponseAction(logger));
       });
-    g_lua.writeFunction("newRemoteLogger", [client](const std::string& remote) {
-        return std::make_shared<RemoteLogger>(ComboAddress(remote));
+    g_lua.writeFunction("newRemoteLogger", [client](const std::string& remote, boost::optional<uint16_t> timeout, boost::optional<uint64_t> maxQueuedEntries, boost::optional<uint8_t> reconnectWaitTime) {
+        return std::make_shared<RemoteLogger>(ComboAddress(remote), timeout ? *timeout : 2, maxQueuedEntries ? *maxQueuedEntries : 100, reconnectWaitTime ? *reconnectWaitTime : 1);
       });
 
 }
