+++ pdns/dnsdistdist/dnsdist-protobuf.cc
@@ -1,0 +1,1 @@
+../dnsdist-protobuf.cc
