+++ pdns/dnsdistdist/dnsdist-protobuf.hh
@@ -1,0 +1,1 @@
+../dnsdist-protobuf.hh
