--- pdns/dnsdistdist/dnsdist-remotelogger.cc
@@ -1,1 +1,0 @@
-../dnsdist-remotelogger.cc
