--- pdns/dnsdistdist/dnsdist-remotelogger.hh
@@ -1,1 +1,0 @@
-../dnsdist-remotelogger.hh
