+++ pdns/dnsdistdist/remote_logger.cc
@@ -1,0 +1,1 @@
+../remote_logger.cc
