+++ pdns/dnsdistdist/remote_logger.hh
@@ -1,0 +1,1 @@
+../remote_logger.hh
